requirements.txt
streamlit==1.29.0
transformers==4.38.2
sentence-transformers==2.2.2
chromadb==0.4.22
PyPDF2==3.0.1
torch==2.2.1
openai==1.6.0
python-dotenv==1.0.1
uuid==1.30